Rename file to match class name and extract other classes out of it for readability
